+++ TractorMod/release-notes.md
@@ -10,6 +10,7 @@
 * Fixed incorrect speed changes when a different speed buff expires while riding the tractor (via Stardew Valley 1.4 change).
 * Fixed performance when using the hoe attachment when surrounded by untillable tiles.
 * Fixed issue where a dismissed tractor would return to the previous garage position if the garage was moved that day.
+* Fixed error when using tools as custom attachments to attack monsters.
 * Improved translations. Thanks to Hesperusrus (updated Russian)!
 
 ## 4.8.4